requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiofiles==22.1.0; python_version >= '3.7' and python_version < '4.0'
alembic==1.8.1
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
cryptography==38.0.3; python_version >= '3.6'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
greenlet==2.0.0.post0
httptools==0.5.0; python_full_version >= '3.5.0'
mako==1.2.3; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
multidict==6.0.2; python_version >= '3.7'
passlib[bcrypt]==1.7.4
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
python-jose[cryptography]==3.3.0
pyyaml==6.0; python_version >= '3.6'
rsa==4.9; python_version >= '3.6' and python_version < '4.0'
sanic-ext==22.9.0
sanic-routing==22.8.0
sanic==22.9.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.43
typing-extensions==4.4.0; python_version >= '3.7'
ujson==5.5.0; sys_platform != 'win32' and implementation_name == 'cpython'
uvloop==0.17.0; sys_platform != 'win32' and implementation_name == 'cpython'
websockets==10.4; python_version >= '3.7'
